requirements.txt
web3==4.8.3
py-evm==0.3.0a1
py-solc-x==1.1.1
z3-solver==4.8.5.0
psutil==5.7.2
networkx==2.5.1
pydot==1.4.2
slither-analyzer==0.8.3
matplotlib==3.3.4
loguru==0.7.2
pandas==1.5.2
seaborn==0.12.0
crytic-compile==0.2.4
solc-select==0.2.0
